chore(core): separate unit and integration tests for webmangler/manglers
