+++ packages/core/src/manglers/__tests__/unit/index.test.ts
@@ -1,0 +1,29 @@
+import { expect } from "chai";
+
+import * as manglers from "../../index";
+
+suite("The webmangler/manglers exports", function() {
+  test("has BuiltInManglers", function() {
+    expect(manglers).to.have.property("BuiltInManglers");
+  });
+
+  test("has RecommendedManglers", function() {
+    expect(manglers).to.have.property("RecommendedManglers");
+  });
+
+  test("has CssClassMangler", function() {
+    expect(manglers).to.have.property("CssClassMangler");
+  });
+
+  test("has CssVariableMangler", function() {
+    expect(manglers).to.have.property("CssVariableMangler");
+  });
+
+  test("has HtmlAttributeMangler", function() {
+    expect(manglers).to.have.property("HtmlAttributeMangler");
+  });
+
+  test("has HtmlIdMangler", function() {
+    expect(manglers).to.have.property("HtmlIdMangler");
+  });
+});
