+++ packages/core/src/manglers/builtin/__tests__/common/values.ts
@@ -1,6 +1,6 @@
 import type { TestValuesSets } from "@webmangler/testing";
 
-import type { RecommendedManglersOptions } from "../types";
+import type { BuiltInManglersOptions } from "../../types";
 
 const booleanOptional = [
   undefined,
@@ -28,6 +28,7 @@ export const optionsValues = {
   disableCssClassMangling: booleanOptional,
   disableCssVarMangling: booleanOptional,
   disableHtmlAttrMangling: booleanOptional,
+  disableHtmlIdMangling: booleanOptional,
 
   classAttributes: stringsOptional,
   classNamePattern: stringOrStringsOptional,
@@ -41,4 +42,10 @@ export const optionsValues = {
   attrNamePattern: stringOrStringsOptional,
   keepAttrPrefix: stringOptional,
   reservedAttrNames: stringsOptional,
-} as TestValuesSets<keyof RecommendedManglersOptions>;
+
+  idAttributes: stringsOptional,
+  idNamePattern: stringOrStringsOptional,
+  keepIdPrefix: stringOptional,
+  reservedIds: stringsOptional,
+  urlAttributes: stringsOptional,
+} as TestValuesSets<keyof BuiltInManglersOptions>;
