+++ packages/core/src/manglers/recommended/__tests__/common/index.ts
@@ -1,0 +1,5 @@
+import { optionsValues } from "./values";
+
+export {
+  optionsValues,
+};
